Separating python and cpp files
