--- program.py
@@ -1,22 +1,0 @@
-import sys
-import cudaq
-
-print(f"Running on target {cudaq.get_target().name}")
-qubit_count = int(sys.argv[1]) if 1 < len(sys.argv) else 2
-
-# Define our kernel.
-@cudaq.kernel
-def kernel():
-    # Allocate our qubits.
-    qvector = cudaq.qvector(qubit_count)
-    # Place the first qubit in the superposition state.
-    h(qvector[0])
-    # Loop through the allocated qubits and apply controlled-X,
-    # or CNOT, operations between them.
-    for qubit in range(qubit_count - 1):
-        x.ctrl(qvector[qubit], qvector[qubit + 1])
-    # Measure the qubits.
-    mz(qvector)
-
-result = cudaq.sample(kernel)
-print(result)
