+++ python/times_out/classical/12-16 classical.txt
@@ -1,0 +1,5 @@
+12 20066132.8
+13 42148398.5
+14 185188681.7
+15 1389243950.9
+16 3970450933.7
