--- times_out/classical/12-14 classical.txt
@@ -1,3 +1,0 @@
-12 60999933.5
-13 136492401.4
-14 99900838.0
