--- times_out/classical/12-16 classical.txt
@@ -1,5 +1,0 @@
-12 85913586.9
-13 48362226.6
-14 142604804.8
-15 456683875.5
-16 1337319275.0
